current working version

+++ rebuild.sh
@@ -2,8 +2,10 @@
 
 sudo echo "rebuilding system"
 
-sudo echo "removing *.nixbak files"
+if [ "$1" != "-s" ]; then
+    sudo echo "removing *.nixbak files"
 
-sudo find ~/ -name "*.nixbak" -type f -delete
+    sudo find ~/ -name "*.nixbak" -type f -delete
+fi
 
 sudo nixos-rebuild switch -p solitude --flake ./#serenity --log-format internal-json -v |& nom --json
